Put changed methods in a seperate file and other small changes
